backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.7
PyMuPDF==1.23.8
python-docx==0.8.11
reportlab==4.0.7
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.2 